requirements.txt
python-nmap
rich
python-dotenv
google-generativeai
